apps/overlay_manager/requirements.txt
requests==2.27.1
python-dotenv==0.19.2
timeloop==1.0.2
